+++ opyoid/bindings/class_binding/class_binding_to_provider_adapter.py
@@ -1,7 +1,7 @@
 from opyoid.bindings.binding import Binding
 from opyoid.bindings.binding_to_provider_adapter import BindingToProviderAdapter
 from opyoid.bindings.registered_binding import RegisteredBinding
-from opyoid.injection_state import InjectionState
+from opyoid.injection_context import InjectionContext
 from opyoid.provider import Provider
 from opyoid.target import Target
 from opyoid.typings import InjectedT
@@ -12,9 +12,12 @@ from .class_binding import ClassBinding
 class ClassBindingToProviderAdapter(BindingToProviderAdapter[ClassBinding]):
     """Creates a Provider from an ClassBinding."""
 
-    def accept(self, binding: Binding[InjectedT], state: InjectionState) -> bool:
+    def accept(self, binding: Binding[InjectedT], context: InjectionContext[InjectedT]) -> bool:
         return isinstance(binding, ClassBinding)
 
-    def create(self, binding: RegisteredBinding[ClassBinding[InjectedT]], state: InjectionState) -> Provider:
+    def create(self,
+               binding: RegisteredBinding[ClassBinding[InjectedT]],
+               context: InjectionContext[InjectedT]) -> Provider:
         new_target = Target(binding.raw_binding.bound_type, binding.raw_binding.annotation)
-        return state.provider_creator.get_provider(new_target, state)
+        new_context = context.get_child_context(new_target)
+        return new_context.get_provider()