+++ opyoid/bindings/instance_binding/instance_binding_to_provider_adapter.py
@@ -1,7 +1,7 @@
 from opyoid.bindings.binding import Binding
 from opyoid.bindings.binding_to_provider_adapter import BindingToProviderAdapter
 from opyoid.bindings.registered_binding import RegisteredBinding
-from opyoid.injection_state import InjectionState
+from opyoid.injection_context import InjectionContext
 from opyoid.provider import Provider
 from opyoid.typings import InjectedT
 from .from_instance_provider import FromInstanceProvider
@@ -12,9 +12,10 @@ from .instance_binding import InstanceBinding
 class InstanceBindingToProviderAdapter(BindingToProviderAdapter[InstanceBinding]):
     """Creates a Provider from an InstanceBinding."""
 
-    def accept(self, binding: Binding[InjectedT], state: InjectionState) -> bool:
+    def accept(self, binding: Binding[InjectedT], context: InjectionContext[InjectedT]) -> bool:
         return isinstance(binding, InstanceBinding)
 
     def create(self,
-               binding: RegisteredBinding[InstanceBinding[InjectedT]], state: InjectionState) -> Provider[InjectedT]:
+               binding: RegisteredBinding[InstanceBinding[InjectedT]],
+               context: InjectionContext[InjectedT]) -> Provider[InjectedT]:
         return FromInstanceProvider(binding.raw_binding.bound_instance)