+++ opyoid/providers/providers_factories/from_binding_provider_factory.py
@@ -1,8 +1,7 @@
 import logging
 
-from opyoid.injection_state import InjectionState
+from opyoid.injection_context import InjectionContext
 from opyoid.provider import Provider
-from opyoid.target import Target
 from opyoid.typings import InjectedT
 from .from_registered_binding_provider_factory import FromRegisteredBindingProviderFactory
 from .provider_factory import ProviderFactory
@@ -16,15 +15,16 @@ class FromBindingProviderFactory(ProviderFactory):
     def __init__(self) -> None:
         self._from_registered_binding_provider_factory = FromRegisteredBindingProviderFactory()
 
-    def accept(self, target: Target[InjectedT], state: InjectionState) -> bool:
-        while target not in state.binding_registry:
-            if not state.parent_state:
+    def accept(self, context: InjectionContext[InjectedT]) -> bool:
+        while not context.has_binding():
+            if not context.injection_state.parent_state:
                 return False
-            state = state.parent_state
+            context = context.get_new_state_context(context.injection_state.parent_state)
         return True
 
-    def create(self, target: Target[InjectedT], state: InjectionState) -> Provider[InjectedT]:
-        if target not in state.binding_registry:
-            return state.parent_state.provider_creator.get_provider(target, state.parent_state)
-        binding = state.binding_registry.get_binding(target)
-        return self._from_registered_binding_provider_factory.create(binding, state)
+    def create(self, context: InjectionContext[InjectedT]) -> Provider[InjectedT]:
+        if not context.has_binding():
+            new_context = context.get_new_state_context(context.injection_state.parent_state)
+            return new_context.get_provider()
+        binding = context.get_binding()
+        return self._from_registered_binding_provider_factory.create(binding, context)