+++ opyoid/providers/providers_factories/jit_provider_factory.py
@@ -1,19 +1,19 @@
-from opyoid.bindings import SelfBinding, SelfBindingToProviderAdapter
-from opyoid.injection_state import InjectionState
+from opyoid.bindings import RegisteredBinding, SelfBinding, SelfBindingToProviderAdapter
+from opyoid.injection_context import InjectionContext
 from opyoid.provider import Provider
-from opyoid.target import Target
 from opyoid.typings import EMPTY, InjectedT
 from .provider_factory import ProviderFactory
-from ...bindings.registered_binding import RegisteredBinding
 
 
 class JitProviderFactory(ProviderFactory):
     def __init__(self):
-        self.provider_factory = SelfBindingToProviderAdapter()
+        self._provider_factory = SelfBindingToProviderAdapter()
 
-    def accept(self, target: Target[InjectedT], state: InjectionState) -> bool:
-        return state.options.auto_bindings and target.default is EMPTY and not isinstance(target.type, str)
+    def accept(self, context: InjectionContext[InjectedT]) -> bool:
+        return context.injection_state.options.auto_bindings \
+               and context.target.default is EMPTY \
+               and not isinstance(context.target.type, str)
 
-    def create(self, target: Target[InjectedT], state: InjectionState) -> Provider[InjectedT]:
-        return self.provider_factory.create(
-            RegisteredBinding(SelfBinding(target.type, annotation=target.annotation)), state)
+    def create(self, context: InjectionContext[InjectedT]) -> Provider[InjectedT]:
+        return self._provider_factory.create(
+            RegisteredBinding(SelfBinding(context.target.type, annotation=context.target.annotation)), context)