+++ opyoid/providers/providers_factories/optional_provider_factory.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from opyoid.injection_state import InjectionState
+from opyoid.injection_context import InjectionContext
 from opyoid.provider import Provider
 from opyoid.target import Target
 from opyoid.type_checker import TypeChecker
@@ -11,11 +11,10 @@ from .provider_factory import ProviderFactory
 class OptionalProviderFactory(ProviderFactory):
     """Returns the Provider for an optional type target."""
 
-    def accept(self, target: Target[InjectedT], state: InjectionState) -> bool:
-        return TypeChecker.is_optional(target.type)
+    def accept(self, context: InjectionContext[InjectedT]) -> bool:
+        return TypeChecker.is_optional(context.target.type)
 
-    def create(self,
-               target: Target[Optional[InjectedT]],
-               state: InjectionState) -> Provider[InjectedT]:
-        new_target = Target(target.type.__args__[0], target.annotation)
-        return state.provider_creator.get_provider(new_target, state)
+    def create(self, context: InjectionContext[Optional[InjectedT]]) -> Provider[InjectedT]:
+        new_target = Target(context.target.type.__args__[0], context.target.annotation)
+        new_context = context.get_child_context(new_target)
+        return new_context.get_provider()