+++ opyoid/providers/providers_factories/provider_provider_factory.py
@@ -1,5 +1,5 @@
 from opyoid.bindings import FromInstanceProvider
-from opyoid.injection_state import InjectionState
+from opyoid.injection_context import InjectionContext
 from opyoid.provider import Provider
 from opyoid.target import Target
 from opyoid.type_checker import TypeChecker
@@ -10,11 +10,10 @@ from .provider_factory import ProviderFactory
 class ProviderProviderFactory(ProviderFactory):
     """Returns the provider for a provider target by transforming a Provider into a FromInstanceProvider."""
 
-    def accept(self, target: Target[InjectedT], state: InjectionState) -> bool:
-        return TypeChecker.is_provider(target.type)
+    def accept(self, context: InjectionContext[InjectedT]) -> bool:
+        return TypeChecker.is_provider(context.target.type)
 
-    def create(self,
-               target: Target[Provider[InjectedT]],
-               state: InjectionState) -> Provider[Provider[InjectedT]]:
-        new_target = Target(target.type.__args__[0], target.annotation)
-        return FromInstanceProvider(state.provider_creator.get_provider(new_target, state))
+    def create(self, context: InjectionContext[Provider[InjectedT]]) -> Provider[Provider[InjectedT]]:
+        new_target = Target(context.target.type.__args__[0], context.target.annotation)
+        new_context = context.get_child_context(new_target)
+        return FromInstanceProvider(new_context.get_provider())