+++ opyoid/providers/providers_factories/tuple_provider_factory.py
@@ -1,7 +1,7 @@
 from typing import List, Tuple
 
 from opyoid.bindings import FromClassProvider
-from opyoid.injection_state import InjectionState
+from opyoid.injection_context import InjectionContext
 from opyoid.provider import Provider
 from opyoid.target import Target
 from opyoid.type_checker import TypeChecker
@@ -12,11 +12,10 @@ from .provider_factory import ProviderFactory
 class TupleProviderFactory(ProviderFactory):
     """Creates a Provider that groups the target tuple items providers."""
 
-    def accept(self, target: Target[InjectedT], state: InjectionState) -> bool:
-        return TypeChecker.is_tuple(target.type)
+    def accept(self, context: InjectionContext[InjectedT]) -> bool:
+        return TypeChecker.is_tuple(context.target.type)
 
-    def create(self,
-               target: Target[Tuple[InjectedT]],
-               state: InjectionState) -> Provider[Tuple[InjectedT]]:
-        new_target = Target(List[target.type.__args__[0]], target.annotation)
-        return FromClassProvider(tuple, [state.provider_creator.get_provider(new_target, state)], None, {})
+    def create(self, context: InjectionContext[Tuple[InjectedT]]) -> Provider[Tuple[InjectedT]]:
+        new_target = Target(List[context.target.type.__args__[0]], context.target.annotation)
+        new_context = context.get_child_context(new_target)
+        return FromClassProvider(tuple, [new_context.get_provider()], None, {})