+++ tests/test_bindings/test_instance_binding/test_instance_binding_to_provider_adapter.py
@@ -31,8 +31,8 @@ class TestInstanceBindingToProviderAdapter(unittest.TestCase):
         self.assertFalse(self.adapter.accept(ProviderBinding(MyType, create_autospec(Provider)), self.state))
 
     def test_create_returns_provider(self):
-        provider = self.adapter.create(
-            RegisteredBinding(InstanceBinding(MyType, self.instance)), self.providers_creator)
+        provider = self.adapter.create(RegisteredBinding(InstanceBinding(MyType, self.instance)),
+                                       self.providers_creator)
 
         instance = provider.get()
         self.assertIs(instance, self.instance)